HBASE-22694 Fallback to hbase.zookeeper.quorum if fs.hboss.sync.zk.connectionString is undefined
Simple code change, but some refactoring to add testing for the change.

EmbeddedZK was made to be an object which we instantiate, rather than
static state and static methods. This lets us re-use the same code for
the contract tests without stomping on one-another.

Also adds in .gitignore entries for Eclipse metadata.

Closes #6

Signed-off-by: Sean Busbey <busbey@apache.org>